Sheetmind/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Settings & environment
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Supabase
supabase==2.9.1

# Redis
redis==5.1.1

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.27.2
email-validator>=2.0.0

# AI (OpenRouter via OpenAI-compatible client)
openai>=1.58.1,<2.0.0

# LangChain (Agent framework) — pin to 0.3.x to avoid backtracking
langchain>=0.3.0,<1.0.0
langchain-openai>=0.3.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-community>=0.3.0,<1.0.0

# RAG (Vector search) — uses API-based embeddings (no PyTorch needed)
chromadb>=0.5.0,<1.0.0

# Payments
dodopayments>=1.66.0,<2.0.0

# API Analytics
api-analytics[fastapi]>=1.2.0

# Utilities
python-multipart==0.0.12
